requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
jinja2==3.1.6
python-multipart==0.0.20
SQLAlchemy==2.0.43
pydantic==2.11.7
pydantic-settings==2.10.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
cryptography==45.0.6
httpx==0.28.1
numpy==2.3.2
pandas==2.3.2
scikit-learn==1.7.1
ccxt==4.4.95
MetaTrader5==5.0.4874; platform_system == "Windows"
